Add simple update checker
Added a simple update checker that notifies the user, on the
settings-tab, if a new version has been released.

The update checker disregards pre-releases and drafts, and does not
properly parse semver-versionnumbers that contains extra pre-release info
or build-metadata. This project wont use those things in releases, so they
are fine to fail on.