Add option to download albums into separate folders.
Closes #5

For Reddit-downloads, albums may be saved into folders under the target
folder. The name of each album-folder is the name of the reddit-post.

For Local-filtering, this option preserves the file hierarchy of the
original folder in the target folder rather than just dumping the files
in the root of the target folder.

Imgur-albums cannot contain nested albums so this option is not supported
for Imgur-downloads.