Add top-level exception handler, unwrap AggregateException. Closes #6
Sometimes WebExceptions seem to be thrown as-is, while other times they
are wrapped in an AggregateException. Not entirely sure why, but now we
catch wrapped WebExceptions as well.

Additionally, with a new top-level exception handler the program
shouldn't outright crash anymore if an unhandled error occurs. Instead
it provides a pop-up that lets the user retrieve the error-message and
stacktrace that can then be included in a github issue.